+++ mayan/apps/mayan_statistics/__init__.py
@@ -1,3 +1,5 @@
 from __future__ import unicode_literals
 
+from .classes import StatisticLineChart, StatisticNamespace
+
 default_app_config = 'mayan_statistics.apps.StatisticsApp'